+++ diive/pkgs/gapfilling/__init__.py
@@ -8,6 +8,7 @@ Feature engineering pipeline and model scoring included.
 Part of the diive library: https://github.com/holukas/diive
 """
 
+from diive.core.ml.results import GapFillingResult
 from diive.pkgs.gapfilling.interpolate import linear_interpolation
 from diive.pkgs.gapfilling.mds import _FluxMDS
 from diive.pkgs.gapfilling.randomforest_ts import RandomForestTS
@@ -16,6 +17,7 @@ from diive.pkgs.gapfilling.scores import prediction_scores
 from diive.pkgs.gapfilling.longterm import LongTermGapFillingRandomForestTS, LongTermGapFillingXGBoostTS
 
 __all__ = [
+    'GapFillingResult',
     'linear_interpolation',
     '_FluxMDS',
     'RandomForestTS',